pyproject.toml
[project]
name = "paracelsus"
version = "0.1.0"
description = "AWS ELT Pipeline POC with LocalStack, Meltano, and dbt"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
dev = [
    "ruff>=0.8.3",
    "basedpyright>=1.14.0",
    "pre-commit>=3.6.0",
]

# Dependencies for synthetic data generation (used in Docker)
synthetic = [
    "faker>=28.0.0",
    "psycopg2-binary>=2.9.9",
    "boto3>=1.34.0",
    "openpyxl>=3.1.2",
    "pandas>=2.2.0",
]

# Dependencies for mock HubSpot API (used in Docker)
mock-api = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "faker>=28.0.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.basedpyright]
typeCheckingMode = "basic"
exclude = [
    ".venv",
    "mock_hubspot",      # Docker-only, deps installed in container
    "synthetic_data",    # Docker-only, deps installed in container
    "terraform/lambda",  # Lambda functions with minimal deps
]
